requirements.txt
python-magic~=0.4.18
Pillow~=8.0.1